v1.5 enhanced program logic
+ deal with different type of BT task (single file, folder with single
  episode, folder with multiple episode)
+ enhanced logic for detect episode and anime name
+ enhanced logic for detect share files
+ support SP episode and versions

+++ DL_Stop.sh
@@ -41,19 +41,23 @@ function NORMAL {
 
 function BT_SINGLE {
 	echo [DL_Stop.sh]" "[Single BT Mode]
-	CLEAN_FILES
 	echo [DL_Stop.sh]" "moving file...
-	mv -v "$path" "$downloadpath/TEMP"
-	if [ "$?" != "0" ]; then rm -rfv "$path"; fi
+	if [[ -f "$path.upload" ]]; then
+		mv -v "$path" "$downloadpath/TEMP"
+		if [ "$?" != "0" ]; then rm -rfv "$path"; fi
+	fi
+	CLEAN_FILES
 	exit 0
 }
 
 function BT_FOLDER {
 	echo [DL_Stop.sh]" "[Multi BT Mode]
-	CLEAN_FILES
 	echo [DL_Stop.sh]" "moving file...
-	mv -v "$path" "$downloadpath/TEMP"
-	if [ "$?" != "0" ]; then rm -rfv "$path"; fi
+	if [[ -f "$path.upload" ]]; then
+		mv -v "$path" "$downloadpath/TEMP"
+		if [ "$?" != "0" ]; then rm -rfv "$path"; fi
+	fi
+	CLEAN_FILES
 	exit 0
 }
